bookstore/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.3.0
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
django-filter==23.3
